[Aggregated Metrics] Helper for performantly getting relevant observations for a metric(s) (Recidiviz/recidiviz-data#34725)
## Description of the change

In the refactored, performant version of aggregated metrics queries, the
basic building block aggregated metric query will only build aggregated
metrics for a single observation type (to avoid aggregating across
multiple observation types).

This PR introduces a helper query that I'll use later to select any
observation relevant to a collection of selectors that all reference the
same observation type. If a row matches any of the selectors, it will be
returned.

The tests I added in this PR give a pretty decent show what the
generated queries look like.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#29291

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 77478aba135c5c02bb3582293c23aa47d9023242