[US_AZ] Fix minor bugs in incarceration periods view (Recidiviz/recidiviz-data#34113)
## Description of the change

This PR fixes a few bugs in the incarceration periods ingest view query
that were making the view less accurate.
1. Standardizes location name at the beginning of the query instead of
the end, so the singular location can be carried forward rather than 3
separate fields with different levels of detail.
2. Orders movements correctly when constructing periods. Transfers OUT
come before Returns from Hospital/Court, which come before Transfers IN.

This leaves large discrepancies in internal vs external population
counts for the PHOENIX facility specifically; however, it brings the
accuracy within 3% in both of the facilities we are launching in next
week (TUCSON and PERRYVILLE). I will prioritize figuring out the rest
before the state-wide launch!

Sandbox ingest pipeline
[here](https://console.cloud.google.com/dataflow/jobs/us-central1/2024-11-05_07_47_48-8331697042005180327;graphView=0?project=recidiviz-staging&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22)))

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 366f91e8f9f07e45ee194dabb9ee579d7ded9c56