[US_AZ] Add home plan status indicators (Recidiviz/recidiviz-data#34800)
## Description of the change

This tweaks the home plan preprocessed view so that we can more closely
match what users see in ACIS regarding home plan progress.

If a home plan is in progress and has not been reviewed & returned to a
COIII or CCO Supervisor, then the status will be "Home Plan In
Progress". This is different than what users see in ACIS because the
dropdown menu they choose from in ACIS is only relevant to decisions
about home plans, not their actual statuses.

I loaded this new view to a sandbox at
`recidiviz-staging.elise_analyst_data.us_az_home_plan_preprocessed`.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 549743ebc50fcf2e6482a69e23980aaf21229970