[Schema] Migrate legacy schema definitions to YAML file. (Recidiviz/recidiviz-data#34532)
## Description of the change

In order to do Recidiviz/recidiviz-data#32861, the comments in `schema.py` needed to live
somewhere else. This PR copies those over into a YAML file and then
tests that all entities have field definitions (except for fields noted
in the test).

### Next Steps
- I will update our docs generator in Recidiviz/recidiviz-data#32861
- This will also have some new documentation for sentencing entities 👀
- then document normalized entities in Recidiviz/recidiviz-data#32863.
- We can also then add the descriptions to BigQuery schema generation
and see our descriptions in BigQuery <3

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| **Type: Feature** | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: c790eed91093a6b78c0daf38e260003b6ad4f2ed