[us_tn] add non-null exemptions (Recidiviz/recidiviz-data#34846)
## Description of the change

in the new raw data infra, a column that is all NULLs in a full
historical file will fail to import (i caught a few files that had
errantly all null PKs). in tn, there are quite a few columns that are
all NULLs in all full historical files currently in big query; i
consulted the wise genevieve for advice and she said:
>if they are all null, we probably don't use them and thus that's why we
didn't note anything before so fine to exempt

>please put direct quote in there

>"xoxo, genevieve"

thus, thir PR adds non-null exemptions to columns that have all nulls
for a particular fields to unblock raw data import, with an open issue
Recidiviz/recidiviz-data#34845 that tracks the work of determining if this is expected (i.e.
they are actually all NULLs in tn's database, meaning the database field
is unused) or if there is some sort of issue with their export process
for these columns

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#34590

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 9c74def8bd49d016551f16be0b6bd40471bf68f8