[JII] include data on ineligible residents in US_ME SCCP product view (Recidiviz/recidiviz-data#34852)
Assuming I've done everything right up to this point, there should be no
visible change in either Opportunities or Workflows due to the
additional rows exported by this change. This sets us up to start
displaying more of this data in Opportunities.

In the spirit of cleaning things up I also removed the view that powered
the old paper prototype, since we don't need it anymore and having two
similarly named views in that folder confuses me every single time.

GitOrigin-RevId: 23225a7e5ebb290637a932ed636b03219847647e