[US_AZ] Hydrate resident_record.release_date (Recidiviz/recidiviz-data#34890)
## Description of the change

Hydrates `release_date` for 34,697 of 36,206 rows for AZ in the
`resident_record` view (96%) with the same SED value used in the AZ
resident metadata. I also added a TODO to migrate this to a sentence
sessions v2 table (similar to the v1 projected incarceration completion
date spans) once that is available.

## Related issues

Closes Recidiviz/recidiviz-data#34881
Closes Recidiviz/recidiviz-data#34859

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 786a1c508127f01684e60a8a62a5befa4cc9a516