[ust_tn] remove `InvolvementType` from `IncompatiblePair` (Recidiviz/recidiviz-data#34841)
## Description of the change

the `InvolvementType` field is not present in the header row for
`IncompatiblePair`(and thus are all nulls in the raw data table) for all
`update_datetime`s in big query:
- `2023-08-31T11:14:12:073546`
- `2024-11-04T12:11:17:423384`
- `2024-11-05T12:06:09:945638`
- `2024-11-06T12:09:44:901379`

i did not check if all of the intermediary files (which are not present
in big query due to raw data pruning are also missing these columns).
this file and column do not seem to be used downstream anywhere.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#34590

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 43ce3fa9f78a982c98b02c9e54605bfb4a9a031a