[US_ND] First drafts of PSI views (Recidiviz/recidiviz-data#34427)
## Description of the change

As I was validating that we had the data we need, I found it easiest to
write very rough first drafts of the views we will need for the PSI
project in ND. This PR adds those drafts to a brand new US_ND folder of
sentencing views!

One thing to note is that while we are set up to import the
`docstars_psi` data and are receiving 2-week incremental transfers
nightly, something went wrong with the full historical transfer of the
file, so there is very little data in that table as of now. The raw data
config for that file is also currently only deployed to staging, so
there will be no data in prod until it is deployed there next week!

There is definitely still lots of work to be done to hone these - I left
some comments and helpful hints where I could. Happy to answer any
questions along the way!
## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 59829f99be16d068695ba220a8189b5bf4f9fea9