[US_AZ] Projected Dates View Modification (Recidiviz/recidiviz-data#34999)
## Description of the change

The `us_az_projected_dates` view needed to be modified to ensure that
there are not duplicates within a day for our ACIS and csbd/ercd dates.
There was a qualify statement attempting to do this but was in fact
filtering out the ACIS dates inadvertently. This fixes that bug.

Now we find that there are 28763 more cases that now have an ACIS date
currently that did not have one before.
All validations are passing.
This is loaded to sandbox under `shalin`.
After reviewing some cases that were not showing active dates in the
Recidiviz dashboard, I can confirm they now show the proper dates as
they should as a result of this change.

You can use the following script to test cases seen on dashboard as Fast
Trackers.
```
SELECT
  FIRST_NAME,
  SURNAME,
  pd.*,
  *
FROM
  `recidiviz-staging.workflows_views.us_az_approaching_acis_or_recidiviz_tpr_request_record_materialized` wv
  LEFT JOIN `recidiviz-staging.us_az_raw_data_up_to_date_views.PERSON_latest` person
  ON person.PERSON_ID = wv.external_id
  INNER JOIN `recidiviz-staging.normalized_state.state_person_external_id` pei
  ON pei.external_id = person.PERSON_ID
  LEFT JOIN `recidiviz-staging.shalin_analyst_data.us_az_projected_dates_materialized` pd
  on pd.person_id = pei.person_id
  WHERE id_type = 'US_AZ_PERSON_ID'
  AND ADC_NUMBER = '...'
order by pd.start_date
```

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 6534c53d2fb9befd023fedb04d14ffe11ad86909